Bugfix for ReactionAirSea
Add parameters
    atm_partial_pressure_min_zero (default false)
    ocean_conc_min_zero (default false)

to limit partial pressure or concentration to > 0.0
when calculating fluxes.

Previous behaviour was to limit ocean concentration to > 0.0, but only
when isotopes not in use.

This was inconsistent, and wrong in some cases eg if a configuration
explicitly allowed -ve ocean O2 to represent unresolved CH4/H2S etc.